Implement autocomplete from within a data table, expanding beyond the table's borders
